requirements_dev.txt
# Generic python dependencies
black==25.1.0
mypy==1.17.0
pip==25.1.1
pre-commit==4.2.0
ruff==0.12.4

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.262
syrupy>=4.6.1

# Component dependencies
hassil>=2.0.4
home-assistant-intents>=2025.3.28
mutagen>=1.47.0
ha-ffmpeg>=3.2.0
pymicro-vad>=1.0.1
pyspeex-noise>=1.0.2
PyTurboJPEG==1.8.0

openai==1.97.0
voluptuous_openapi>=0.0.5
audioop-lts>=0.2.1
